release 0.0.21-beta source code for go

+++ services/mpc/v1/model/model_delete_template_request.go
@@ -15,7 +15,7 @@ import (
 
 // Request Object
 type DeleteTemplateRequest struct {
-	TemplateId int64 `json:"template_id"`
+	TemplateId int32 `json:"template_id"`
 }
 
 func (o DeleteTemplateRequest) String() string {
